requirements.txt
langchain
chromadb
openai
python-dotenv
unstructured
tiktoken